requirements.txt
geopandas>=0.14.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.6.0
rasterio>=1.3.0
folium>=0.14.0
requests>=2.31.0
psycopg2-binary>=2.9.0
matplotlib>=3.7.0
cartopy>=0.22.0
contextily>=1.4.0
osmnx>=1.6.0
overpy>=0.7.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0.0
click>=8.1.0
tqdm>=4.65.0
python-dotenv>=1.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.4.0
